requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
docusign-esign==4.0.0
email_validator==2.2.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
logtail-python==0.3.2
msgpack==1.1.0
nose==1.3.7
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
